requirements.txt
asgiref==3.3.1
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
flake8==3.8.4
isort==5.7.0
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==2.0.1
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
